src/test/java/baekjoon/_2893Test.java
package baekjoon;

import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import java.io.*;

import static org.assertj.core.api.Assertions.assertThat;

class _2893Test {
    private final ByteArrayOutputStream output = new ByteArrayOutputStream();
    private _2893 object;

    @BeforeEach
    void setUpOutputStream() {
        System.setOut(new PrintStream(output));
    }
    @AfterEach
    void resotreOutputStream() {
        System.setOut(System.out);
        output.reset();
    }

    void setInputStream(String input) {
        InputStream in = new ByteArrayInputStream(input.getBytes());
        System.setIn(in);
    }

    @Test
    void testCase1() throws IOException {
        // given
        setInputStream("18");

        String compareResult = "4";

        // when
        object = new _2893();
        object.solution();

        // then
        assertThat(output.toString().trim())
                .isEqualTo(compareResult);
    }

    @Test
    void testCase2() throws IOException {
        // given
        setInputStream("4");

        String compareResult = "-1";

        // when
        object = new _2893();
        object.solution();

        // then
        assertThat(output.toString().trim())
                .isEqualTo(compareResult);
    }

    @Test
    void testCase3() throws IOException {
        // given
        setInputStream("6");

        String compareResult = "2";

        // when
        object = new _2893();
        object.solution();

        // then
        assertThat(output.toString().trim())
                .isEqualTo(compareResult);
    }

    @Test
    void testCase4() throws IOException {
        // given
        setInputStream("9");

        String compareResult = "3";

        // when
        object = new _2893();
        object.solution();

        // then
        assertThat(output.toString().trim())
                .isEqualTo(compareResult);
    }

    @Test
    void testCase5() throws IOException {
        // given
        setInputStream("11");

        String compareResult = "3";

        // when
        object = new _2893();
        object.solution();

        // then
        assertThat(output.toString().trim())
                .isEqualTo(compareResult);
    }

}